Comentando e arrumando codigo do content_extractor.py
