+++ .travis.yml
@@ -1,7 +1,6 @@
 # command to run tests
 language: python
 python:
-  - "3.5"
   - "3.6"
   - "3.7"
   - "3.8"
