+++ requirements.txt
@@ -5,3 +5,5 @@ pytest==5.3.5
 Sphinx==2.3.1
 sphinxcontrib-blockdiag==2.0.0
 PyMuPDF==1.16.13
+flashtext==2.7.0
+pdf2image==1.12.1
